Add dual disentangle model implementation with dataset and training scripts
- Introduced new dataset class `DualDisentangleDataset` for paired modality data handling.
- Implemented `EBMJEPALoss` for loss calculation in the dual model training.
- Created model building functions to wrap unimodal models and set up cross-modal predictors.
- Added configuration files for various training setups including 3D and 2D data types.
- Developed main training script to manage the training loop and logging.
- Included visualization functions to log original and reshaped data spaces to W&B.
- Added SLURM scripts for training on different GPU partitions.

+++ multimodal_experiments/ssl_dual_alignment/vis.py
@@ -41,8 +41,8 @@ def plot_dual_geometry_reshaping_view(dual_model, data_a, data_b, param_values, 
     """Plots 4-way view: Input A -> Output A -> Output B -> Input B."""
     dual_model.eval()
     with torch.no_grad():
-        output_a, _ = dual_model.model_a(torch.tensor(data_a, device=device, dtype=torch.float64))
-        output_b, _ = dual_model.model_b(torch.tensor(data_b, device=device, dtype=torch.float64))
+        output_a, _ = dual_model.model_a(torch.tensor(data_a, device=device, dtype=torch.float32))
+        output_b, _ = dual_model.model_b(torch.tensor(data_b, device=device, dtype=torch.float32))
 
     output_a = output_a.detach().cpu().numpy()
     output_b = output_b.detach().cpu().numpy()
